Initial import of Findanator.
At this stage, there are two forms, one to view alarms, another to add/edit
alarms. There is also a database adapter, so all form data is saved to
database. The UI is very minimal, with TextView and EditText used for
all data entry. This needs to be drastically improved upon to use a tab
interface, time picker, etc.