Clicking on row of alarm now directs to edit page.
Previously, you had to click on the name of the alarm alone to
navigate to the EditAlarm page. This has been fixed to allow
you to click anywhere on the row. This was done by changing
the android:layout_width property of alarm row from
wrap_content to fill_parent.